requirements.txt
Flask
gunicorn
requests
python-dotenv
psycopg2-binary